Factor out response data processing to responses.rs
